refactor: Installer階層を2段化し、Domain構築とScope組み立てを静的ヘルパーへ切り出す
#137で導入したInstallerBaseは対象を「描画を持つUIバックエンド」に限定していたため、
CommandLineInstallerに約127行の重複が残り、_themeが基底にあることがCLI版と共通化できる
範囲を狭めていた(#145)。

- InstallerBase を全Installer共通の骨格へ縮小
  - _options を共通フィールドへ引き上げ(既定値の差し替えはprotected Options経由、null時の
    フォールバックはCreateFallbackOptions()のoverride)
  - Install()の骨格は Domain構築 → コマンド登録 → BuildBackend() → Scope組み立て
  - 唯一の抽象点を BuildBackend(options, in DomainContext) : BackendContext にした
- RenderingInstallerBase を新設(描画を持つバックエンド共通)
  - _animation・入力生成・Coordinator構築を保持し、BuildBackend()をsealed overrideで実装
  - 派生の実装点は BuildRenderingContext() のみ
- 静的ヘルパーへ切り出し
  - DomainBuilder … Domain構築とコマンド登録
  - ScopeBuilder … Componentsの分配とScope組み立て、失敗時のCleanUp
  - ThemeSync … ColorPaletteAccessor/CursorFlashSpeedAccessorへのテーマ同期
- DomainContext / BackendContext を public トップレベル型として Context/ へ配置
- _theme を基底から外し、ImmediateModeInstaller / UIToolkitInstaller が各自で宣言
- CommandLineInstaller を InstallerBase 配下へ(247→70行)

具象Installerの型名・名前空間・アセンブリ、および各シリアライズフィールド名は不変。
EditMode 253件 / PlayMode 67件ともPassed。
