+++ Assets/YukimaruGames/Terminal/Runtime/Composition/Building.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 255a2f98a258f4ba4ac43ff18f5890dc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
