+++ Assets/YukimaruGames/Terminal/Runtime/Composition/Building/ScopeBuilder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0f26a5461ff174d58a8adcd62d74cf6d
