+++ Assets/YukimaruGames/Terminal/Runtime/Composition/Building/ThemeSync.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6f84281df77e34be1b5291396ff0ca78
